Changed request processing logic & Added migration

+++ CatalogManagementService/src/Application/Processors/RemoveProductRequestProcessor.cs
@@ -5,12 +5,18 @@ using Core.DTO;
 
 namespace CatalogManagementService.Application;
 
-public class RemoveProductRequestProcessor(IProductRepository repository) : IRequestProcessor<ProductDTO, Product>
+public class RemoveProductRequestProcessor(IProductRepository repository, ILogger<RemoveProductRequestProcessor> logger) 
+    : IRequestProcessor<Guid, Product>
 {
-    public async Task<Product> Process(ProductDTO data)
+    public async Task<Product> Process(Guid data)
     {
-        var product = data.ToDomain();
+        var product = await repository.GetAsync(data);
+        if (product is null)
+            throw new InvalidOperationException($"DELETE: Product with id '{data}' not found.");
+
         await repository.DeleteAsync(product);
+        
+        logger.LogInformation($"Product with id '{product.Id}' removed.");
         return product;
     }
 }
